+++ src-tauri/src/lib.rs
@@ -151,6 +151,7 @@ pub fn run() {
             toggle_logs::toggle_logs,
             commands::import_library,
             commands::get_tracks,
+            commands::get_track,
             commands::get_play_queue,
             commands::set_play_queue,
             commands::get_global_tags,